requirements.txt
tqdm
pandas
opendatasets
tensorflow[and-cuda]
tkinter
customtkinter
opencv-python
pillow